subscribe: restore rotation-seam convergence in cold replay
The readable-log refactor replaced WalkFromCursor's #190 seam machinery
with a single sealed-sweep + active-sweep pass, on the premise that the
StopSeq floor bound closed the rotation seam. It does not: the manifest
read and the ActiveIndex read are non-atomic, so a rotation completing
between them (rotateLocked publishes segment N, then bumps activeIdx to
N+1) leaves N reachable via neither source in one pass. The single-pass
walk then let ColdReader.Read jump the cursor to the floor, silently
dropping N's durable events — an egress data-loss regression.

A second, related defect: walkActiveRegion's emitOne checked
ev.Seq >= StopSeq before the hole case ev.Seq > current, so a gap whose
next active event sat at/above the floor stopped with no error and was
silently skipped rather than surfaced.

Restore the bounded convergence loop the design doc prescribes
(specs/notes/2026-07-06-readable-log-implementation-design.md): sealed
sweep -> active sweep -> if still below StopSeq, re-sweep the sealed
region. walkActiveRegion now stops cleanly at a below-floor hole with
current unchanged (never jumps a gap). Publish-before-bump gives a
happens-before that makes exactly one no-progress pass expected at a live
seam; a second consecutive no-progress pass is an invariant violation and
fails loud rather than spinning or skipping. Add an OnSeamRetry hook for
deterministic testing and a future seam-hit metric.

Tests: strengthen the concurrent-rotation test to assert exact
[start, floor) completeness (contiguity + ceiling alone could not catch
the seam skip); add a deterministic seam-convergence test that publishes
a withheld segment mid-walk via OnSeamRetry; update the fail-loud gap
test's error string. Red-first verified: a single-pass variant fails the
seam test.

Also delete dead Entry.approxBytes (hot-ring fallout) and fix two
branch-local lint errors (unconvert in steady.go, errcheck in
readlog_test.go) that were blocking CI.

ROAST_SKIP=1: roast on these files already run this session (2 HIGH
findings surfaced, both fixed), re-roasted clean with cross-model
verification the quick-mode gate doesn't run.

+++ internal/ingest/orchestrator/steady.go
@@ -109,7 +109,7 @@ func (o *Orchestrator) runSteadyState(ctx context.Context) error {
 			TombstoneCap:          tombstoneCap,
 			CompactionTrigger:     o.compactionTrigger,
 			SegmentMetrics:        o.cfg.SegmentMetrics,
-			ReadLogRetentionBytes: int64(o.cfg.ReadLogRetentionBytes),
+			ReadLogRetentionBytes: o.cfg.ReadLogRetentionBytes,
 			OnEvent:               onEvent,
 			OnAfterSeal:           o.cfg.IngestOnAfterSeal,
 			ReconnectBackoff:      o.cfg.LiveReconnectBackoff,
